+++ src/Controller/Blockexplorer/BlockController.php
@@ -269,7 +269,7 @@ class BlockController extends ApiController
      *      @SWG\Parameter(
      *          name="hideConnections",
      *          in="query",
-     *          type="bool",
+     *          type="boolean",
      *          description="The field used to hide connect transactions"
      *      ),
      *     @SWG\Parameter(