+++ src/Controller/Blockexplorer/MessageController.php
@@ -178,7 +178,7 @@ class MessageController extends ApiController
      *      @SWG\Parameter(
      *          name="hideConnections",
      *          in="query",
-     *          type="bool",
+     *          type="boolean",
      *          description="The field used to hide connect transactions"
      *      ),
      *     @SWG\Parameter(