+++ src/Controller/Blockexplorer/NodeController.php
@@ -350,7 +350,7 @@ class NodeController extends ApiController
      *      @SWG\Parameter(
      *          name="hideConnections",
      *          in="query",
-     *          type="bool",
+     *          type="boolean",
      *          description="The field used to hide connect transactions"
      *      ),
      *     @SWG\Parameter(